Extract tests for isFloat64Array to own file
